IRQN Macro added. (create IRQ callbacks)

+++ kernel/pic.h
@@ -6,7 +6,7 @@
 #define IRQ_BASE  0x20
 
 void setup_pic();
-void irq_enable(int irq);
+void irq_enable(uint8_t irq);
 void pic_acknowledge(uint8_t irq);
 
 #endif
